requirements.txt
python-dotenv
requests
pandas
streamlit
pymongo
millify
rich